chore: split remaining shared type buckets
