+++ packages/kernel/src/application/generators/sql-generator.ts
@@ -1,4 +1,4 @@
-import type { CompiledSchemaSql, MetaIndex, MetaSchema, MetaTable, SnapshotRelation } from "../../core/types/shared.types";
+import type { CompiledSchemaSql, MetaIndex, MetaSchema, MetaTable, SnapshotRelation } from "../../core/interfaces";
 import { validateSchema } from "../../core/utils";
 import { createTableSql, quoteIdentifier } from "../../utils/sql-utils";
