+++ packages/kernel/src/core/interfaces/index.ts
@@ -1,1 +1,3 @@
+export * from "./kernel.interface";
 export * from "./meta-kernel.interface";
+export * from "./migration-safety.interface";
