+++ packages/kernel/src/core/interfaces/kernel.interface.ts
@@ -1,10 +1,16 @@
-export type ViewExpression =
-  | string
-  | number
-  | boolean
-  | null
-  | ViewExpression[]
-  | { [key: string]: ViewExpression };
+import type {
+  ApiRouteMethod,
+  ApiRouteOperation,
+  DataScopeType,
+  MergeStrategy,
+  MetaDefinitionKind,
+  MigrationAction,
+  NodeDefinition,
+  PermissionRuleEffect,
+  PermissionRuleEnforcement,
+  Primitive,
+  ViewExpression
+} from "../types";
 
 export interface ViewDefinition {
   name: string;
@@ -13,10 +19,6 @@ export interface ViewDefinition {
   output: OutputDefinition;
   submit?: SubmitDefinition;
 }
-
-export type NodeDefinition = QueryNodeDefinition | MutationNodeDefinition | TransformNodeDefinition | MergeNodeDefinition;
-
-export type MergeStrategy = "objectMerge" | "arrayConcat" | "custom";
 
 export interface BaseNodeDefinition {
   type: "query" | "mutation" | "transform" | "merge";
@@ -67,13 +69,6 @@ export interface DatasourceDefinition {
   description?: string;
 }
 
-export type DataScopeType =
-  | "SELF"
-  | "DEPT"
-  | "DEPT_AND_CHILDREN"
-  | "CUSTOM_ORG_SET"
-  | "TENANT_ALL";
-
 export interface PermissionPolicy {
   id: string;
   resource: string;
@@ -81,8 +76,6 @@ export interface PermissionPolicy {
   roles: string[];
   scope?: DataScopeType;
 }
-
-export type Primitive = string | number | boolean | null;
 
 export interface MetaField {
   name: string;
@@ -153,33 +146,6 @@ export interface SnapshotV1 {
   checksum: string;
 }
 
-export type MigrationAction =
-  | {
-      type: "create_table";
-      table: MetaTable;
-    }
-  | {
-      type: "drop_table";
-      table: MetaTable;
-    }
-  | {
-      type: "add_column";
-      table: string;
-      column: MetaField;
-    }
-  | {
-      type: "drop_column";
-      table: string;
-      column: MetaField;
-    }
-  | {
-      type: "alter_column";
-      table: string;
-      columnName: string;
-      fromType: string;
-      toType: string;
-    };
-
 export interface MigrationDslV1 {
   version: string;
   actions: MigrationAction[];
@@ -197,9 +163,6 @@ export interface CompiledSchemaSql {
   relations: string[];
   statements: string[];
 }
-
-export type ApiRouteOperation = "query" | "mutation";
-export type ApiRouteMethod = "POST";
 
 export interface CompiledApiRouteTarget {
   method: ApiRouteMethod;
@@ -221,9 +184,6 @@ export interface CompiledApiRouteManifest {
   source: "meta-schema";
   routes: CompiledApiRoute[];
 }
-
-export type PermissionRuleEffect = "allow";
-export type PermissionRuleEnforcement = "rbac";
 
 export interface CompiledPermissionRule {
   id: string;
@@ -283,8 +243,6 @@ export interface MigrationAuditRecord {
   requestId: string;
 }
 
-export type MetaDefinitionKind = "view" | "datasource" | "permissionPolicy";
-
 export interface MetaDefinitionByKind {
   view: ViewDefinition;
   datasource: DatasourceDefinition;
@@ -325,6 +283,3 @@ export interface MetaDefinitionDiff {
   toVersion: number;
   changedPaths: string[];
 }
-
-export type LatestMetaDefinitionVersion<K extends MetaDefinitionKind = MetaDefinitionKind> =
-  MetaDefinitionVersion<K>;
