+++ packages/kernel/src/core/interfaces/migration-safety.interface.ts
