+++ packages/kernel/src/utils/sql-utils.ts
@@ -1,4 +1,4 @@
-import type { MetaField } from "../core/types/shared.types";
+import type { MetaField } from "../core/interfaces";
 
 export function quoteIdentifier(value: string): string {
   if (!/^[a-zA-Z_][a-zA-Z0-9_]*$/.test(value)) {
