+++ packages/permission/src/core/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from "./permission.interface";
