+++ packages/permission/src/core/types/index.ts
@@ -1,1 +1,1 @@
-export * from "./shared.types";
+export type * from "./permission.type";
