+++ packages/permission/src/core/types/permission.type.ts
@@ -1,0 +1,6 @@
+export type DataScopeType =
+  | "SELF"
+  | "DEPT"
+  | "DEPT_AND_CHILDREN"
+  | "CUSTOM_ORG_SET"
+  | "TENANT_ALL";
