+++ packages/query/src/core/index.ts
@@ -1,1 +1,2 @@
 export * from "./types";
+export * from "./interfaces";
