+++ packages/query/src/core/types/index.ts
@@ -1,1 +1,1 @@
-export * from "./shared.types";
+export type * from "./query.type";
